requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.3
django-cors-headers==3.11.0
django-rest-auth==0.9.5
django-saml2-auth==2.2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
djangosaml2==1.4.0
drf-jwt==1.19.2
elementpath==2.5.0
idna==3.3
isodate==0.6.1
lxml==4.6.5
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pysaml2==7.1.2
python-dateutil==2.8.2
python3-saml==1.14.0
python3-saml-django==1.2.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
xmlschema==1.10.0
xmlsec==1.3.12
pip install django-phonenumber-field==6.1.0
